+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Details about this file’s format at <http://keepachangelog.com/>. The change log is parsed automatically when minting releases through Fastlane, see `Fastlane/Fastfile`.
 
 ## [Unreleased]
+- Apple Watch app 
 
 ## [1.6.0, build 49] - 2023-04-14Z
 - Widget, UK localization, fixed issue with 12 hours rotation